requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp @ file:///C:/ci/aiohttp_1646806572557/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
albumentations==1.3.0
astor==0.8.1
astunparse==1.6.3
async-timeout @ file:///C:/b/abs_43ozhz2a8g/croots/recipe/async-timeout_1664876362767/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi @ file:///C:/Windows/Temp/abs_6808y9x40v/croots/recipe/cffi_1659598653989/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///C:/ci/click_1646038595831/work
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
contourpy==1.0.5
cryptography @ file:///C:/ci/cryptography_1652083563162/work
cycler==0.11.0
dlib @ file:///D:/bld/dlib_1652035331399/work
docopt==0.6.2
executing==1.1.1
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==2.0.7
fonttools==4.37.4
frozenlist @ file:///C:/ci/frozenlist_1637767271796/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
google-auth==1.35.0
google-auth-oauthlib==0.4.1
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///C:/ci/grpcio_1637590978642/work
h5py @ file:///C:/ci/h5py_1659089875384/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.22.2
importlib-metadata @ file:///C:/ci/importlib-metadata_1648562621412/work
joblib==1.2.0
keras==2.7.0
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.4.4
libclang==14.0.6
Markdown @ file:///C:/ci/markdown_1614364082838/work
matplotlib==3.6.1
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
multidict @ file:///C:/b/abs_6cx_8w3cv2/croot/multidict_1665674238352/work
networkx==2.8.7
numpy @ file:///C:/Windows/Temp/abs_2a1e1vbeag/croots/recipe/numpy_and_numpy_base_1659432712056/work
oauthlib @ file:///C:/b/abs_2eoymqc2ow/croot/oauthlib_1665490906043/work
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging==21.3
pandas==1.5.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
protobuf==3.17.2
pure-eval==0.2.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///C:/ci/pyjwt_1657511236979/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests @ file:///C:/ci/requests_1657735342357/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
tensorboard @ file:///tmp/build/80754af9/tensorboard_1633093581375/work/tensorboard-2.6.0-py3-none-any.whl
tensorboard-data-server @ file:///C:/ci/tensorboard-data-server_1633035225228/work/tensorboard_data_server-0.6.0-py3-none-any.whl
tensorboard-plugin-wit @ file:///C:/tf/b/tensorboard-plugin-wit_1660162132996/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow==2.7.0
tensorflow-determinism==0.3.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3 @ file:///C:/Windows/TEMP/abs_65ynz4fdmi/croots/recipe/urllib3_1659110473919/work
wcwidth==0.2.5
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt @ file:///C:/Windows/Temp/abs_7c3dd407-1390-477a-b542-fd15df6a24085_diwiza/croots/recipe/wrapt_1657814452175/work
yarg==0.1.9
yarl @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_e5nlunspj6/croots/recipe/yarl_1661437086516/work
zipp @ file:///C:/ci/zipp_1652273994994/work
